requirements.txt
# Core dependencies
google-generativeai==0.3.2
pinecone-client==2.2.2
pandas==2.1.1
numpy==1.24.3

# Speech recognition and text-to-speech
SpeechRecognition==3.10.0
pyaudio==0.2.13
edge-tts==6.1.9
pygame==2.5.2

# Translation
googletrans==4.0.0-rc1

# Audio playback
playsound==1.3.0

# Environment management
python-dotenv==1.0.0